Changed name from p2p_kernel to p2p_cpu_kernel
