+++ CMakeLists.txt
@@ -128,9 +128,8 @@ set(source_files ${source_files}
     src/multipole_interactions/multipole_cpu_kernel.cpp
     src/monopole_interactions/cuda_p2p_interaction_interface.cpp
     src/monopole_interactions/p2p_cuda_kernel.cu
+    src/monopole_interactions/p2p_cpu_kernel.cpp
     src/monopole_interactions/p2p_interaction_interface.cpp
-    src/monopole_interactions/p2p_kernel.cpp
-    src/monopole_interactions/p2p_kernel_blocked_interaction.cpp
     src/monopole_interactions/calculate_stencil.cpp
     src/monopole_interactions/p2m_interaction_interface.cpp
     src/monopole_interactions/p2m_kernel.cpp
@@ -180,7 +179,7 @@ set(header_files
     src/monopole_interactions/cuda_p2p_interaction_interface.hpp
     src/monopole_interactions/p2p_interaction_interface.hpp
     src/monopole_interactions/p2p_cuda_kernel.hpp
-    src/monopole_interactions/p2p_kernel.hpp
+    src/monopole_interactions/p2p_cpu_kernel.hpp
     src/monopole_interactions/calculate_stencil.hpp
     src/monopole_interactions/p2m_interaction_interface.hpp
     src/monopole_interactions/p2m_kernel.hpp
