+++ src/monopole_interactions/cuda_p2p_interaction_interface.cpp
@@ -37,14 +37,14 @@ namespace fmm {
                     cudaMemcpyHostToDevice);
 
                 // Launch kernel and queue copying of results
-                const dim3 grid_spec(3, 1, 1);
+                const dim3 grid_spec(3);
                 const dim3 threads_per_block(8, 8, 8);
                 void* args[] = {&(env.device_local_monopoles), &(env.device_blocked_monopoles),
                                 &(env.device_stencil), &(env.device_four_constants),&theta, &dx};
                 gpu_interface.execute(
                     &cuda_p2p_interactions_kernel, grid_spec, threads_per_block, args, 0);
                 void* sum_args[] = {&(env.device_blocked_monopoles)};
-                const dim3 sum_spec(1, 1, 1);
+                const dim3 sum_spec(1);
                 const dim3 threads(512);
                 gpu_interface.execute(
                     &cuda_add_pot_blocks, sum_spec, threads, sum_args, 0);
